# docs/openapi.yaml
openapi: 3.1.0
info:
  title: mKey daemon API
  version: 1.0.0
  description: |
    HTTP API демона mKey (SPEC §9, модуль internal/api).

    Транспорты:
    - Unix-сокет `$XDG_RUNTIME_DIR/mkey/mkey.sock` (права 0600) — для CLI, без токена;
    - `http://127.0.0.1:17420` — для веб-интерфейса: заголовок `Host` только 127.0.0.1/localhost,
      для изменяющих запросов `Origin` — только страницы mKey, токен из `$XDG_RUNTIME_DIR/mkey/token`
      в заголовке `Authorization: Bearer …` или cookie `mkey_session` (ставится при входе по `/?t=<токен>`).

    Сообщения ошибок переводятся на язык из `Accept-Language` (ru, en).
servers:
  - url: http://127.0.0.1:17420
security:
  - bearer: []
  - cookie: []
paths:
...
  /api/v1/projects/{id}/{action}:
    post:
      summary: Включить или выключить проект (action = enable | disable)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: action, in: path, required: true, schema: { enum: [enable, disable] } }
      responses: { "200": { description: Выполнено }, "400": { $ref: "#/components/responses/Error" } }
...
  /api/v1/projects/{id}/export:
    get:
      summary: "Файл проекта для сохранения и обмена (ADR-0027): в раздел devices дописаны имена кнопок устройств, которые встречаются в проекте; комментарии сохраняются; нет проекта — 404"
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "200": { description: "application/yaml (Content-Disposition)" }, "404": { $ref: "#/components/responses/Error" } }
  /api/v1/projects/{id}:
    get:
      summary: "Проект — структура и текст файла ({project, raw, path, error})"
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "200": { description: Проект }, "404": { $ref: "#/components/responses/Error" } }
    put:
      summary: "Сохранить проект после полной проверки — {project} (из конструктора) или {raw} (текст YAML, комментарии сохраняются)"
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "200": { description: Сохранено }, "400": { $ref: "#/components/responses/Error" } }
    delete:
      summary: Удалить проект (файл)
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "200": { description: Удалено }, "400": { $ref: "#/components/responses/Error" } }
  /api/v1/projects/validate:
    post:
      summary: "Проверить проект ({project} или {raw}), не сохраняя"
      responses: { "200": { description: Ошибок нет }, "400": { $ref: "#/components/responses/Error" } }
...
  /api/v1/capture/key:
    post:
      summary: "Ждать нажатия ({timeout_ms, combo}); combo — сочетание до первого отпускания (\"Ctrl+H\")"
      responses:
        "200": { description: "{name, code, kernel, device, device_name}" }
        "408": { $ref: "#/components/responses/Error" }
...
  /api/v1/dsl/to_actions:
    post:
      summary: "Разобрать макрос на отдельные действия (блоки) — {text} → {actions}"
      responses: { "200": { description: "{actions: [{type, value}]}" }, "400": { $ref: "#/components/responses/Error" } }
  /api/v1/actions/to_dsl:
    post:
      summary: "Склеить действия в один макрос — {actions} → {text}; 400, если действие нельзя записать макросом"
      responses: { "200": { description: "{text}" }, "400": { $ref: "#/components/responses/Error" } }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
    cookie: { type: apiKey, in: cookie, name: mkey_session }
  responses:
    Error:
      description: Ошибка в едином формате
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code: { type: string, example: dsl.unknown_key_hint }
                  message: { type: string, example: "неизвестная клавиша «Mous0» — возможно, вы имели в виду «Mouse0»?" }
                  details:
                    type: object
                    properties:
                      pos:
                        type: object
                        properties:
                          line: { type: integer }
                          col: { type: integer }
                      args: { type: object, additionalProperties: { type: string } }
...